+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [1.7.9] - 2020-02-01
+[feature] select message in hex/traffic viewer
+[feature] hide widgets after clearing messages
 ### [1.7.8] - 2020-01-28
 [feature] looking raw traffic in channel/devices messages
 [fix] hex viewer styles